Licensed under GPL version 3.
The default license is GPLv3, for use in personal projects, demos,
research, and GPL-compatible projects.

If you wish to use this code in a project under another software
license, or in an ecosystem (such as the Apple App Store) that is
incompatible with 3rd-party GPL code, then contact me to receive
the LHVM source code under a different license.